Windows: Separate terminal related code from libvisca.h
