+++ visca/libvisca_win32.c
@@ -46,8 +46,7 @@ unsigned int _VISCA_send_packet_with_reply(VISCAInterface_t *iface, VISCACamera_
 
 uint32_t _VISCA_write_packet_data(VISCAInterface_t *iface, VISCACamera_t *camera, VISCAPacket_t *packet)
 {
-	DWORD iBytesWritten;
-	WriteFile(iface->port_fd, packet->bytes, packet->length, &iBytesWritten, NULL);
+	DWORD iBytesWritten = iface->callback->write(iface, packet->bytes, packet->length);
 	if (iBytesWritten < packet->length)
 		return VISCA_FAILURE;
 	else
@@ -58,9 +57,6 @@ uint32_t _VISCA_send_packet(VISCAInterface_t *iface, VISCACamera_t *camera, VISC
 {
 	DWORD iBytesWritten;
 	static long bytesTogether = 0;
-	BOOL rVal = 0;
-	DWORD errors;
-	COMSTAT stat;
 	int nTrials;
 
 	// check data:
@@ -82,10 +78,10 @@ uint32_t _VISCA_send_packet(VISCAInterface_t *iface, VISCACamera_t *camera, VISC
 	// append footer
 	_VISCA_append_byte(packet, VISCA_TERMINATOR);
 
-	for (nTrials = 0; nTrials < 3 && rVal == 0; nTrials++) {
-		if (nTrials > 0)
-			ClearCommError(iface->port_fd, &errors, &stat);
-		rVal = WriteFile(iface->port_fd, &packet->bytes, packet->length, &iBytesWritten, NULL);
+	for (nTrials = 0; nTrials < 3 && iBytesWritten != packet->length; nTrials++) {
+		if (nTrials > 0 && iface->callback->clear_error)
+			iface->callback->clear_error(iface);
+		iBytesWritten = iface->callback->write(iface, &packet->bytes, packet->length);
 	}
 
 	if (iBytesWritten < packet->length) {
@@ -100,12 +96,11 @@ uint32_t _VISCA_send_packet(VISCAInterface_t *iface, VISCACamera_t *camera, VISC
 uint32_t _VISCA_get_packet(VISCAInterface_t *iface)
 {
 	int pos = 0;
-	BOOL rc;
-	DWORD iBytesRead;
+	int iBytesRead;
 
 	// wait for message
-	rc = ReadFile(iface->port_fd, iface->ibuf, 1, &iBytesRead, NULL);
-	if (!rc || iBytesRead == 0) {
+	iBytesRead = iface->callback->read(iface, iface->ibuf, 1);
+	if (iBytesRead <= 0) {
 		// Obtain the error code
 		//m_lLastError = ::GetLastError();
 		_RPTF0(_CRT_WARN, "ReadFile failed.\n");
@@ -118,8 +113,8 @@ uint32_t _VISCA_get_packet(VISCAInterface_t *iface)
 			_RPTF0(_CRT_WARN, "illegal reply packet.\n");
 			return VISCA_FAILURE;
 		}
-		rc = ReadFile(iface->port_fd, iface->ibuf + pos, 1, &iBytesRead, NULL);
-		if (!rc || iBytesRead == 0) {
+		iBytesRead = iface->callback->read(iface, iface->ibuf + pos, 1);
+		if (iBytesRead <= 0) {
 			// Obtain the error code
 			//m_lLastError = ::GetLastError();
 			_RPTF0(_CRT_WARN, "ReadFile failed.\n");
@@ -134,99 +129,6 @@ uint32_t _VISCA_get_packet(VISCAInterface_t *iface)
 /***********************************/
 /*       SYSTEM  FUNCTIONS         */
 /***********************************/
-
-uint32_t VISCA_open_serial(VISCAInterface_t *iface, const char *device_name)
-{
-	BOOL m_bPortReady;
-	HANDLE m_hCom;
-	DCB m_dcb;
-	COMMTIMEOUTS cto;
-
-	m_hCom = CreateFile(device_name, GENERIC_READ | GENERIC_WRITE,
-			    0,    // exclusive access
-			    NULL, // no security
-			    OPEN_EXISTING,
-			    0,     // no overlapped I/O
-			    NULL); // null template
-
-	// Check the returned handle for INVALID_HANDLE_VALUE and then set the buffer sizes.
-	if (m_hCom == INVALID_HANDLE_VALUE) {
-		_RPTF1(_CRT_WARN, "cannot open serial device %s\n", device_name);
-		iface->port_fd = NULL;
-		return VISCA_FAILURE;
-	}
-
-	m_bPortReady = SetupComm(m_hCom, 4, 4); // set buffer sizes
-
-	// Port settings are specified in a Data Communication Block (DCB). The easiest way to initialize a DCB is to call GetCommState to fill in its default values, override the values that you want to change and then call SetCommState to set the values.
-	m_bPortReady = GetCommState(m_hCom, &m_dcb);
-	m_dcb.BaudRate = 9600;
-	m_dcb.ByteSize = 8;
-	m_dcb.Parity = NOPARITY;
-	m_dcb.StopBits = ONESTOPBIT;
-	m_dcb.fAbortOnError = TRUE;
-
-	// =========================================
-	// (jd) added to get a complete setup...
-	m_dcb.fOutxCtsFlow = FALSE;              // Disable CTS monitoring
-	m_dcb.fOutxDsrFlow = FALSE;              // Disable DSR monitoring
-	m_dcb.fDtrControl = DTR_CONTROL_DISABLE; // Disable DTR monitoring
-	m_dcb.fOutX = FALSE;                     // Disable XON/XOFF for transmission
-	m_dcb.fInX = FALSE;                      // Disable XON/XOFF for receiving
-	m_dcb.fRtsControl = RTS_CONTROL_DISABLE; // Disable RTS (Ready To Send)
-	m_dcb.fBinary = TRUE;                    // muss immer "TRUE" sein!
-	m_dcb.fErrorChar = FALSE;
-	m_dcb.fNull = FALSE;
-	// =========================================
-
-	m_bPortReady = SetCommState(m_hCom, &m_dcb);
-
-	// =========================================
-	// (jd) set timeout
-	if (!GetCommTimeouts(m_hCom, &cto)) {
-		// Obtain the error code
-		//m_lLastError = ::GetLastError();
-		_RPTF0(_CRT_WARN, "unable to obtain timeout information\n");
-		CloseHandle(m_hCom);
-		return VISCA_FAILURE;
-	}
-	cto.ReadIntervalTimeout = 100;       /* 20ms would be good, but 100 are for usb-rs232 */
-	cto.ReadTotalTimeoutConstant = 2000; /* 2s  */
-	cto.ReadTotalTimeoutMultiplier = 50; /* 50ms for each char */
-	cto.WriteTotalTimeoutMultiplier = 500;
-	cto.WriteTotalTimeoutConstant = 1000;
-	if (!SetCommTimeouts(m_hCom, &cto)) {
-		// Obtain the error code
-		//m_lLastError = ::GetLastError();
-		_RPTF0(_CRT_WARN, "unable to setup timeout information\n");
-		CloseHandle(m_hCom);
-		return VISCA_FAILURE;
-	}
-	// =========================================
-
-	// If all of these API's were successful then the port is ready for use.
-	iface->port_fd = m_hCom;
-	iface->address = 0;
-
-	return VISCA_SUCCESS;
-}
-
-uint32_t VISCA_unread_bytes(VISCAInterface_t *iface, unsigned char *buffer, uint32_t *buffer_size)
-{
-	// TODO
-	*buffer_size = 0;
-	return VISCA_SUCCESS;
-}
-
-uint32_t VISCA_close_serial(VISCAInterface_t *iface)
-{
-	if (iface->port_fd != NULL) {
-		CloseHandle(iface->port_fd);
-		iface->port_fd = NULL;
-		return VISCA_SUCCESS;
-	} else
-		return VISCA_FAILURE;
-}
 
 uint32_t VISCA_usleep(uint32_t useconds)
 {
